refactor: removed handlecontent
